+++ src/main/java/de/johni0702/minecraft/bobby/BobbyConfig.java
@@ -8,6 +8,8 @@ public class BobbyConfig {
     public static final BobbyConfig DEFAULT = new BobbyConfig();
 
     private boolean enabled = true;
+    @Comment("Do not load block entities (e.g. chests) in fake chunks.\nThese need updating every tick which can add up.\n\nEnabled by default because the render distance for block entities is usually smaller than the server-view distance anyway.")
+    private boolean noBlockEntities = true;
     @Comment("Delays the unloading of chunks which are outside your view distance.\nSaves you from having to reload all chunks when leaving the area for a short moment (e.g. cut scenes).\nDoes not work across dimensions.")
     private int unloadDelaySecs = 60;
     @Comment("Changes the maximum value configurable for Render Distance.\n\nRequires Sodium.")
@@ -19,14 +21,20 @@ public class BobbyConfig {
 
     public BobbyConfig(
             boolean enabled,
+            boolean noBlockEntities,
             int unloadDelaySecs,
             int maxRenderDistance,
             int viewDistanceOverwrite
     ) {
         this.enabled = enabled;
+        this.noBlockEntities = noBlockEntities;
         this.unloadDelaySecs = unloadDelaySecs;
         this.maxRenderDistance = maxRenderDistance;
         this.viewDistanceOverwrite = viewDistanceOverwrite;
+    }
+
+    public boolean isNoBlockEntities() {
+        return noBlockEntities;
     }
 
     public boolean isEnabled() {